requirements.txt
Flask==3.0.0
Pillow==10.2.0
torch==2.1.0
torchvision==0.16.0
Werkzeug==3.0.1

# Core Framework
streamlit==1.31.0

# Computer Vision & Image Processing
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
numpy==1.26.3

# Deep Learning & ONNX Runtime
onnxruntime==1.17.0

# Database
mysql-connector-python==8.3.0

# Password Hashing
bcrypt==4.1.2

# Utilities
python-dateutil==2.8.2


